+++ CHANGELOG.rst
@@ -1,6 +1,15 @@
 Changelog
 =========
 
+1.4.0 (2023-03-20)
+------------------
+* Adjusted the `SS_GROUPBOX` stylesheet: Groupbox title is now vertically
+aligned to be centered with the top line, regardless of the Qt lib. Nothing
+changes for PyQt5/PySide2, and fixes PyQt6/PySide6.
+* Added function `e8()` to calculate the pixel width that is necessary to fit
+`N` number of '8's in the supplied `font`. In analogy to the 'em' font-metric,
+but now focused on displaying numeric values.
+
 1.3.0 (2023-02-27)
 ------------------
 * Neater support for PyQt5, PyQt6, PySide2 and PySide6